Managed to find and fix a few bugs in the View class.
The video metadata file seems now to be processed almost
correctly. We are still missing stuff to sort MP4s first,
regardless of what the "selected" flag in the metadata
file says, an adaption of the template to reflect the
playable videos, and a style sheet to format it all.

Also missing: A flash player, and MUCH MORE TESTING!!!