Attempt to improve the test suite. Especially, there are now more browser tests.
One of them is disabled because I have to figure out why the
registry change in the test suite's setUp() method does not
work: The fspath is not being set to /var/tmp/something, but
instead to /tmp, which causes the test to fail. If I step
through the validator which is being tested (validateFilename()
in the video.py module) and re-adjust the path to the temp
directory which is being generated by the setUp() method, the
test passes. :|